+++ lib/reusable_switch_btn.dart
@@ -1,5 +1,3 @@
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'app_switch.dart';
+export 'app_switch_card.dart';
+export 'normal_switch.dart';